deposito obra usuario

+++ Repositorios/Implementaciones/IRecursosServicio.cs
@@ -6,14 +6,13 @@ namespace Repositorios.Implementaciones
 {
     public interface IRecursosServicio
     {
-        Task<Response<string>>RecursoCargar(RecursosCargarDTO recursosCargarDTO, long depositoId);
+        Task<Response<string>> RecursoCargar(RecursosCargarDTO recursosCargarDTO, long depositoId);
         Task<(bool, string)> RecursosTransladarAdeposito(RecursosTransladarDepositoDTO RecursosTransladarDepositoDTO);
-        Task <Response<List<RecursosPagPrincipalDTO>>> RecursosVerDTO(int empresaId);
+        Task<Response<List<RecursosPagPrincipalDTO>>> RecursosVerDTO(int empresaId);
         Task<(bool, List<RecursosVerDepositoDTO>)> RecursosVerDepositoDTO(int depositoId);
-        //Task<(bool, string)> RecursosActualizarStock(RecursosActualizarDTO recursoActualizarDTO, int depositoId);
         Task<(bool, RecursoStockVerDTO)> ObtenerRecursoPorStockId(int stockId);
         Task<(bool, string)> RecursoEliminarStock(int stockId);
         Task<(bool, object)> VerificarRecursoPorCodigoISO(string codigoISO);
-        Task<(bool, string)> RecursosActualizar(RecursosActualizarDTO recursoActualizarDTO, int recursoId);
+        Task<(bool, string)> RecursosActualizar(RecursosActualizarDTO recursoActualizarDTO, long recursoId);
     }
 }
